roman-numerals: use common tests (plus a few extra tests), add TestVersion, correct limit on example program, use consistent actual/expected order in test, simplify benchmark.
