+++ roman-numerals/cases_test.go
@@ -1,0 +1,31 @@
+package romannumerals
+
+// Source: exercism/x-common
+// Commit: 6985644 Merge pull request #121 from mikeyjcat/add-roman-numerals-test-definition
+
+type romanNumeralTest struct {
+	arabic   int
+	roman    string
+	hasError bool
+}
+
+var romanNumeralTests = []romanNumeralTest{
+	{1, "I", false},
+	{2, "II", false},
+	{3, "III", false},
+	{4, "IV", false},
+	{5, "V", false},
+	{6, "VI", false},
+	{9, "IX", false},
+	{27, "XXVII", false},
+	{48, "XLVIII", false},
+	{59, "LIX", false},
+	{93, "XCIII", false},
+	{141, "CXLI", false},
+	{163, "CLXIII", false},
+	{402, "CDII", false},
+	{575, "DLXXV", false},
+	{911, "CMXI", false},
+	{1024, "MXXIV", false},
+	{3000, "MMM", false},
+}
